rename ggml-cpu-traits
